Service test finished & ProjectResource controller

+++ demo/src/main/java/com/example/demo/Service/impl/ProjectColumnServiceImp.java
@@ -4,6 +4,7 @@ import com.example.demo.Repository.*;
 import com.example.demo.ResourceManager.model.*;
 import com.example.demo.Service.*;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -19,8 +20,14 @@ public class ProjectColumnServiceImp implements ProjectColumnService {
     }
 
     @Override
-    public void deleteColumn(Integer column_id){
+    public Boolean deleteColumn(Integer column_id){
+        ProjectColumns column=repository.findById(column_id).orElse(null);
+        if (column == null) {
+            System.out.println("column not exist");
+            return false;
+        }
         repository.deleteById(column_id);
+        return true;
     }
 
     @Override
